Update FTFSI
Added forcedList=DIO<6>  - allow an array of file names to pull instead of using LISTF to get the file list.
It will only alter the program if you populate this var with the calling program.   I only added this to the jBase section because I don't know enough about the flavors to change them.
 IF forcedList#'' THEN
   LST=forcedList
 END ELSE
   SELECT FTMP TO LST
 END

The only issue is that the program seems to do some additional MD digging that I have not figured out how to stop.  This still has reduced the VS Code load time from 15 to 5 seconds.